Create an initial empty commit for the root
It turns out that git does not define an initial zero commit and this
sometimes causes problems at the boundaries (certain corner cases fail
on the "first commit"). For example, rebase cannot rebase the first
commit which can be trying.

For this reason, we can force git to create an initial, empty commit (by
using `git commit --allow-empty`. This commit will simply sit in the
head of our repo and need never be touched.

Signed-off-by: Saurabh Prakash <saurabhprakash.smvit@gmail.com>